nouvelle version

+++ src/sets/TestPays.java
@@ -12,18 +12,18 @@ public class TestPays {
 		HashSet<Pays> setPays = new HashSet<>();
 
 		// Ajout des pays avec leurs habitants et leurs PIB
-		setPays.add(new Pays("Italie", 60000000, 33000));
-		setPays.add(new Pays("Allemagne", 83000000, 44000));
-		setPays.add(new Pays("USA", 331000000, 56000));
-		setPays.add(new Pays("Inde", 1400000000, 5000));
-		setPays.add(new Pays("UK", 68000000, 37000));
-		setPays.add(new Pays("Japon", 126000000, 35000));
-		setPays.add(new Pays("Chine", 1400000000, 14000));
-		setPays.add(new Pays("Russie", 145000000, 24000));
-		setPays.add(new Pays("France", 67000000, 39000));
+		setPays.add(new Pays("Italie", 60000000, 33000.0));
+		setPays.add(new Pays("Allemagne", 83000000, 44000.0));
+		setPays.add(new Pays("USA", 331000000, 56000.0));
+		setPays.add(new Pays("Inde", 1400000000, 5000.0));
+		setPays.add(new Pays("UK", 68000000, 37000.0));
+		setPays.add(new Pays("Japon", 126000000, 35000.0));
+		setPays.add(new Pays("Chine", 1400000000, 14000.0));
+		setPays.add(new Pays("Russie", 145000000, 24000.0));
+		setPays.add(new Pays("France", 67000000, 39000.0));
 
 		Pays paysAvecMaxPIBparHabitant = null;
-		int maxPIBparHab = Integer.MIN_VALUE;
+		double maxPIBparHab = Double.MIN_VALUE;
 		for (Pays pays : setPays) {
 			if (pays.getPibHab() > maxPIBparHab) {
 				maxPIBparHab = pays.getPibHab();
@@ -35,7 +35,7 @@ public class TestPays {
 		System.out.println(paysAvecMaxPIBparHabitant);
 
 		Pays paysPibTotPlusImp = null;
-		int pibTotPlusImp = Integer.MIN_VALUE;
+		double pibTotPlusImp = Double.MIN_VALUE;
 		for (Pays pays : setPays) {
 			if (pays.getPIBTotal() > pibTotPlusImp) {
 				pibTotPlusImp = pays.getPIBTotal();
@@ -48,9 +48,9 @@ public class TestPays {
 
 		// Recherche du pays avec le PIB total le plus petit
 		Pays paysAvecMinPIBTotal = null;
-		int minPIBTotal = Integer.MAX_VALUE;
+		double minPIBTotal = Double.MAX_VALUE;
 		for (Pays pays : setPays) {
-			int PIBTotal = pays.getPIBTotal();
+			double PIBTotal = pays.getPIBTotal();
 			if (PIBTotal < minPIBTotal) {
 				minPIBTotal = PIBTotal;
 				paysAvecMinPIBTotal = pays;
